--- backend/core/.gitignore
@@ -1,3 +1,0 @@
-.classpath
-.project
-.settings/
