+++ backend/core/.idea/libraries/Maven__io_netty_netty_handler_proxy_4_1_49_Final.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: io.netty:netty-handler-proxy:4.1.49.Final">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/netty/netty-handler-proxy/4.1.49.Final/netty-handler-proxy-4.1.49.Final.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/io/netty/netty-handler-proxy/4.1.49.Final/netty-handler-proxy-4.1.49.Final-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/netty/netty-handler-proxy/4.1.49.Final/netty-handler-proxy-4.1.49.Final-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
