+++ backend/core/.idea/libraries/Maven__io_vertx_vertx_auth_common_4_0_0.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: io.vertx:vertx-auth-common:4.0.0">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/vertx/vertx-auth-common/4.0.0/vertx-auth-common-4.0.0.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/io/vertx/vertx-auth-common/4.0.0/vertx-auth-common-4.0.0-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/vertx/vertx-auth-common/4.0.0/vertx-auth-common-4.0.0-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
