+++ backend/core/pom.xml
@@ -1,8 +1,8 @@
 <project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.amanda</groupId>
-    <artifactId>CookIn</artifactId>
+    <groupId>com.amanda.code.cookin</groupId>
+    <artifactId>cookin</artifactId>
     <version>1.0.0</version>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -10,8 +10,8 @@
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
         <!-- vert.x properties -->
-        <vertx.version>3.9.0</vertx.version>
-
+        <vertx.version>4.0.0</vertx.version>
+        <vertx.verticle>com.amanda.code.CookinService</vertx.verticle>
         <!-- Maven plugins -->
         <vertx-maven-plugin.version>1.0.9</vertx-maven-plugin.version>
     </properties>
@@ -34,21 +34,6 @@
         <dependency>
             <groupId>io.vertx</groupId>
             <artifactId>vertx-web</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-simple</artifactId>
-            <version>1.7.30</version>
-        </dependency>
-        <dependency>
-            <groupId>io.vertx</groupId>
-            <artifactId>vertx-mongo-client</artifactId>
-            <version>4.0.0</version>
-        </dependency>
-        <dependency>
-            <groupId>com.googlecode.json-simple</groupId>
-            <artifactId>json-simple</artifactId>
-             <version>1.1.1</version>
         </dependency>
     </dependencies>
     <build>
