+++ backend/core/src/main/java/com/amanda/code/data/Gender.java
@@ -1,0 +1,6 @@
+package com.amanda.code.data;
+
+public enum Gender {
+    MALE,
+    FEMALE
+}
