+++ backend/core/src/main/java/com/amanda/code/data/Student.java
@@ -1,0 +1,6 @@
+package com.amanda.code.data;
+
+public class Student {
+    String name,regNo,Course;
+
+}
