--- backend/src/main/java/com/amanda/CookIn.java
@@ -1,12 +1,0 @@
-package com.amanda;
-
-import io.vertx.core.AbstractVerticle;
-
-public class CookIn extends AbstractVerticle {
-
-    @Override
-    public void start() {
-
-    }
-
-}
